Refactor health controllers into one
The `DataController` is an example of what we are currently doing. Most of `/health` was migrated from another project. Based on how we have built `DataController` and what we have just done with the other controllers it makes sense to refactor 'health'.
