+++ app/routes/health.routes.js
@@ -1,14 +1,12 @@
 'use strict'
 
-const AirbrakeController = require('../controllers/health/airbrake.controller.js')
-const DatabaseController = require('../controllers/health/database.controller.js')
-const InfoController = require('../controllers/health/info.controller.js')
+const HealthController = require('../controllers/health.controller.js')
 
 const routes = [
   {
     method: 'GET',
     path: '/health/airbrake',
-    handler: AirbrakeController.index,
+    handler: HealthController.airbrake,
     options: {
       description: 'Used by the delivery team to confirm error logging is working correctly in an environment. ' +
         'NOTE. We expect this endpoint to return a 500'
@@ -17,7 +15,7 @@ const routes = [
   {
     method: 'GET',
     path: '/health/database',
-    handler: DatabaseController.index,
+    handler: HealthController.database,
     options: {
       description: 'Used by the delivery team to confirm we can connect to the database. It also returns us some ' +
         'useful stats about each table.'
@@ -26,7 +24,7 @@ const routes = [
   {
     method: 'GET',
     path: '/health/info',
-    handler: InfoController.index,
+    handler: HealthController.info,
     options: {
       description: 'Used by the delivery team to confirm we can connect to our other apps and services. It also ' +
       'returns us the version and commit hash for each one.'
